Fix lost encrypted chats dialog

+++ telegramqml.cpp
@@ -3530,6 +3530,9 @@ void TelegramQml::messagesGetDialogs_slt(qint64 id, qint32 sliceCount, const QLi
     if(p->database) {
         Q_FOREACH(qint64 dId, removedDialogs)
         {
+            if(p->dialogs[dId]->encrypted())
+                continue;
+
             p->database->deleteDialog(dId);
             insertToGarbeges(p->dialogs.value(dId));
         }
